Se creó un record para hacer un DTO en el json de la info que se mostrará con el fin de mostrar al usuario solo la informacion de titulo, año y lugares disponibles de las películas. Se ajustó en la clase Controller, primero el record, luego el método que llama a esa información de nuestro backend Finalmetne se creó en nuestros recursos estáticos: el archivo html film.html y su respectivo javascript film.js, en el html es la página donde se muestra la info que llamaremos con nuestro json, en el javascript importamos añgunos métodos de nuestro util y los ponemos para llenar nuestra infromaación de la película con lo necesario.

+++ src/main/java/be/vdab/cinefest/films/Film.java
@@ -6,10 +6,10 @@ class Film {
     private final long id;
     private final String titel;
     private final int jaar;
-    private final long vrijePlaatsen;
+    private final int vrijePlaatsen;
     private final BigDecimal aankoopprijs;
 
-    public Film(long id, String titel, int jaar, long vrijePlaatsen, BigDecimal aankoopprijs) {
+    public Film(long id, String titel, int jaar, int vrijePlaatsen, BigDecimal aankoopprijs) {
         this.id = id;
         this.titel = titel;
         this.jaar = jaar;
@@ -28,7 +28,7 @@ class Film {
         return jaar;
     }
 
-    public long getVrijePlaatsen() {
+    public int getVrijePlaatsen() {
         return vrijePlaatsen;
     }
